Prototype 1 commit
SharedPrefeences now used.
User is now only able to send a message if their are in a band.
When sending a message the user selects the band to send the message out
from by using a drop-down box which displays the user's bands they are
in.
If the user is not in a band then they only have the option to recieve
messages from the subscriptions they made.