Prototype 1 commit
Bug fixed in login_user.php to prevent a user from logging in to the app
that doesn't exist in the database.
Created log_it.php which appends any messages sent to it in a log file.
Code to log any connections made to the server such as messages sent
from a band or a profile being added to thr database is now logged in
the log file.