Prototype1 commit
28/01/2013 -    Add function in menu to allow user to send feedback to
developer relating to the current activity being viewed.
Need to remember to set permission for a folder on the server to access
it!

Managed to get Menu system working for all versions of Android :-)

Swipe left and right in the Main Activity!!!!! Although not over
buttons.. does this matter?

Bug fixed in SingUp!

Changed MainActivity so that it now uses a XML layout to overcome swipe
listener issues.

05/02/2013 -    Added the first "All" feed.

Make messages appear with band name and picture. Add picture to cache
for quicker loading next time. Used someone elses code.

Made the size of buttons consistent throughout app.

06/02/2013 -  Changed feeds.xml to 'username'.xml so that the user name
login in and out and have the appropriate feedlist used according to the
username currently signed in.

Added Feeds.
http://www.ibm.com/developerworks/xml/library/x-tipmvdom/index.html

http://stackoverflow.com/questions/4903758/android-how-to-refresh-listview-contents

Check to see whether the spinner to slecet which band to send a message
from disappears should the user only be linked to one band profile.
"SendMessages"   view.setVisiblility(View.GONE)

Added Menus to the Feeds.

Swipe now works well!

http://stackoverflow.com/questions/8326599/horizontal-swipe-not-detected-in-scrollviews-parent/8327453#8327453

Check to See if band name is already in use during profile creation (at
StepOne), is now impelemented!

Present search results in a Listview with a pic.

send requests so that user can be approved from menu. User set up tmp
exchange and queue for band.